requirements.txt
matplotlib==3.4.3
networkx==2.6.3
numpy==1.21.2
ogb==1.3.2
PyYAML==6.0
rdkit==2009.Q1-1
scikit_learn==1.0.2
scipy==1.7.1
torch==1.9.0
torch_geometric==2.0.2
torch_scatter==2.0.8
torch_sparse==0.6.12
tqdm==4.62.3
tensorboard==2.7.0
